requirements.txt
asgiref==3.7.2
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
backports.zoneinfo==0.2.1
bleach==5.0.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
constantly==23.10.4
cryptography==42.0.3
daphne==4.1.2
datetime-truncate==1.1.1
defusedxml==0.8.0rc2
diff-match-patch==20230430
Django==4.2.10
django-admincharts==0.4.1
django-bleach==3.1.0
django-bower==5.2.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.7
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-rest-framework==0.1.0
django-schema-graph==2.2.1
django-taggit==5.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==24.11.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inflection==0.5.1
Jinja2==3.1.3
MarkupPy==1.14
MarkupSafe==2.1.5
nose==1.3.7
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pillow==10.2.0
psycopg2==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.21
PyJWT==1.7.1
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python-nvd3==0.14.2
python-slugify==1.1.4
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
service-identity==24.1.0
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
tinycss2==1.1.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.9.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.2.0
webencodings==0.5.1
xlrd==2.0.1
xlwt==1.3.0
zope.interface==6.3
